REFACTOR : DTO 리팩토링
- 협의한 DTO 명명 규칙에 따라 클래스 이름 변경
- 패키지 분리
