+++ src/main/java/com/hansung/hansungcommunity/dto/free/FreeBoardMainDto.java
@@ -1,0 +1,28 @@
+package com.hansung.hansungcommunity.dto.free;
+
+import com.hansung.hansungcommunity.entity.FreeBoard;
+import lombok.Data;
+
+import java.time.LocalDateTime;
+
+@Data
+public class FreeBoardMainDto {
+
+    private Long id;
+    private String title;
+    private String content;
+    private String writer;
+    private LocalDateTime createdDate;
+    private int bookmark;
+//  private int reply; // 댓글 기능 구현 후 추가
+
+    public FreeBoardMainDto(FreeBoard board) {
+        this.id = board.getId();
+        this.title = board.getTitle();
+        this.content = board.getContent();
+        this.writer = board.getUser().getNickname();
+        this.createdDate = board.getCreatedAt();
+        this.bookmark = board.getBookmarks();
+    }
+
+}
