+++ src/main/java/com/hansung/hansungcommunity/dto/free/FreeBoardRequestDto.java
@@ -1,32 +1,31 @@
-package com.hansung.hansungcommunity.dto;
+package com.hansung.hansungcommunity.dto.free;
 
 import com.hansung.hansungcommunity.entity.FreeBoard;
-import com.hansung.hansungcommunity.entity.FreeBoardBookmark;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
 @Data
 @NoArgsConstructor
-public class FreeBoardDto {
+public class FreeBoardRequestDto {
 
     private Long id;
     private String title;
     private String content;
 
-    public FreeBoardDto(FreeBoard board) {
+    public FreeBoardRequestDto(FreeBoard board) {
         this.id = board.getId();
         this.title = board.getTitle();
         this.content = board.getContent();
     }
 
-    public FreeBoardDto(Long id, String title, String content) {
+    public FreeBoardRequestDto(Long id, String title, String content) {
         this.id = id;
         this.title = title;
         this.content =  content;
     }
 
-    public static FreeBoardDto of(FreeBoard freeBoard) {
-        return new FreeBoardDto(
+    public static FreeBoardRequestDto of(FreeBoard freeBoard) {
+        return new FreeBoardRequestDto(
                 freeBoard.getId(),
                 freeBoard.getTitle(),
                 freeBoard.getContent()
