+++ src/main/java/com/hansung/hansungcommunity/dto/qna/QnaBoardMostViewedDto.java
@@ -1,4 +1,4 @@
-package com.hansung.hansungcommunity.dto;
+package com.hansung.hansungcommunity.dto.qna;
 
 import com.hansung.hansungcommunity.entity.QnaBoard;
 import lombok.Data;
@@ -7,7 +7,7 @@ import lombok.Data;
  * 조회수가 높은 게시글 조회를 위한 DTO
  */
 @Data
-public class MostViewedQnaBoardsDto {
+public class QnaBoardMostViewedDto {
 
     private Long id;
     private String title;
@@ -15,7 +15,7 @@ public class MostViewedQnaBoardsDto {
     private String language;
     private int point;
 
-    private MostViewedQnaBoardsDto(Long id, String nickname, String title, String language, int point) {
+    private QnaBoardMostViewedDto(Long id, String nickname, String title, String language, int point) {
         this.id = id;
         this.writer = nickname;
         this.title = title;
@@ -23,8 +23,8 @@ public class MostViewedQnaBoardsDto {
         this.point = point;
     }
 
-    public static MostViewedQnaBoardsDto of(QnaBoard board) {
-        return new MostViewedQnaBoardsDto(
+    public static QnaBoardMostViewedDto of(QnaBoard board) {
+        return new QnaBoardMostViewedDto(
                 board.getId(),
                 board.getUser().getNickname(),
                 board.getTitle(),
