+++ src/main/java/com/hansung/hansungcommunity/dto/recruit/RecruitBoardRequestDto.java
@@ -1,4 +1,4 @@
-package com.hansung.hansungcommunity.dto;
+package com.hansung.hansungcommunity.dto.recruit;
 
 import lombok.Data;
