+++ src/main/java/com/hansung/hansungcommunity/entity/FreeReply.java
@@ -1,11 +1,10 @@
 package com.hansung.hansungcommunity.entity;
 
 
-import com.hansung.hansungcommunity.dto.FreeReplyDto;
+import com.hansung.hansungcommunity.dto.free.FreeReplyDto;
 import lombok.AccessLevel;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
-import org.hibernate.annotations.ColumnDefault;
 
 import javax.persistence.*;
 import java.util.ArrayList;
