+++ src/test/java/com/hansung/hansungcommunity/controller/QnaBoardApiControllerTest.java
@@ -1,7 +1,7 @@
 //package com.hansung.hansungcommunity.controller;
 //
-//import com.hansung.hansungcommunity.dto.QnaBoardDto;
-//import com.hansung.hansungcommunity.dto.UserDto;
+//import com.hansung.hansungcommunity.dto.qna.QnaBoardDto;
+//import com.hansung.hansungcommunity.dto.user.UserDto;
 //import com.hansung.hansungcommunity.entity.QnaBoard;
 //import com.hansung.hansungcommunity.entity.User;
 //import com.hansung.hansungcommunity.repository.QnaBoardRepository;
